requirements.txt
Flask==2.3.0
requests==2.28.2
pyyaml==6.0
qrcode==7.3.1
reportlab==3.6.12
gunicorn==20.1.0
